# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games/{id}/book:
    post:
      operationId: bookSpot
      tags: [bookings]
      summary: Initiate booking - creates Stripe checkout session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookSpotBody"
      responses:
        "200":
          description: Stripe checkout URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Validation error or slot unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /payment/verify:
    get:
      operationId: verifyPayment
      tags: [payment]
      summary: Verify a payment and confirm booking
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
        - name: gameId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
        "400":
          description: Payment not completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Booking:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        userId:
          type: string
        team:
          type: integer
        slotIndex:
          type: integer
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        paymentId:
          type: ["string", "null"]
        bookedAt:
          type: string
          format: date-time
      required:
        - id
        - gameId
        - userId
        - team
        - slotIndex
        - paymentStatus
        - bookedAt
...
    BookSpotBody:
      type: object
      properties:
        team:
          type: integer
        slotIndex:
          type: integer
      required:
        - team
        - slotIndex
    CheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
        sessionId:
          type: string
      required:
        - checkoutUrl
        - sessionId
    PaymentVerifyResponse:
      type: object
      properties:
        status:
          type: string
        booking:
          $ref: "#/components/schemas/Booking"
        gameId:
          type: string
      required:
        - status
        - gameId

    # ==================== DASHBOARD SCHEMAS ====================
...